feat(frontend): consolidate personal profile page into single view (#794)
Replace the Profile/Activity tab switcher and lg:grid-cols-5 split on
/profile with one single-column-first page: an identity+streaks hero,
Profile Details, Badges, and My Engagements all stack vertically at
the page's existing max-w-7xl width. Edit/Save/Cancel move from inline
buttons into the header's quick-actions toolbar via the existing
useEditModeQuickActions hook (AppLayout now wraps pages in
QuickActionsProvider so a non-org page can use it for the first time).
The Share Achievements feature is removed entirely. Legacy ?tab=
bookmarks/redirects scroll to the relevant section instead of
switching tabs. Reword the danger-zone copy to drop the internal
"Keycloak" term.
