+++ backend/tests/VisualTests/AccountTests.cs
@@ -14,8 +14,8 @@ public class AccountTests(AspireFixture fixture) : VisualTestBase(fixture)
 
 		await Page.GotoAsync($"{frontend.GetLeftPart(UriPartial.Authority)}/account");
 
-		// /account redirects to /profile - .First matches the guard already
-		// applied to this same "Edit" button in ProfileOverviewTests.cs.
+		// /account redirects to /profile. Edit is now a single quick-action
+		// button in the header toolbar (#794) - .First is a harmless no-op.
 		var editButton = Page.GetByRole(AriaRole.Button, new() { Name = "Edit" }).First;
 		await Expect(editButton).ToBeVisibleAsync(new() { Timeout = 20_000 });
 		await editButton.ClickAsync();
@@ -34,8 +34,8 @@ public class AccountTests(AspireFixture fixture) : VisualTestBase(fixture)
 
 		await Page.GotoAsync($"{frontend.GetLeftPart(UriPartial.Authority)}/account");
 
-		// /account redirects to /profile - .First matches the guard already
-		// applied to this same "Edit" button in ProfileOverviewTests.cs.
+		// /account redirects to /profile. Edit is now a single quick-action
+		// button in the header toolbar (#794) - .First is a harmless no-op.
 		var editButton = Page.GetByRole(AriaRole.Button, new() { Name = "Edit" }).First;
 		await Expect(editButton).ToBeVisibleAsync(new() { Timeout = 30_000 });
 		await editButton.ClickAsync();
@@ -53,8 +53,8 @@ public class AccountTests(AspireFixture fixture) : VisualTestBase(fixture)
 
 		await Page.GotoAsync($"{frontend.GetLeftPart(UriPartial.Authority)}/account");
 
-		// /account redirects to /profile - .First matches the guard already
-		// applied to this same "Edit" button in ProfileOverviewTests.cs.
+		// /account redirects to /profile. Edit is now a single quick-action
+		// button in the header toolbar (#794) - .First is a harmless no-op.
 		var editButton = Page.GetByRole(AriaRole.Button, new() { Name = "Edit" }).First;
 		await Expect(editButton).ToBeVisibleAsync(new() { Timeout = 20_000 });
 		await editButton.ClickAsync();
